src/main/java/BusStop.java
import java.util.ArrayList;

public class BusStop {

    private String name;
    private ArrayList<Person> queue;

    public BusStop(String name){
        this.name = name;
        this.queue = new ArrayList<Person>();
    }

    public int queueSize() {
        return this.queue.size();
    }


    public void addPassengerToQueue(Person person) {
        this.queue.add(person);
    }

    public void removePassengerFromQueue(BusStop busstop) {
        Person person =
        this.queue.remove(person);

    }
}
